Changed Error_Option to integer
Former-commit-id: 9e0d3b4d2fd91e6b224566ee487e6d7251f48e77
Former-commit-id: 1f4e9f35d8d1f62ed3ac621f58f8d87ed1ca0c55

+++ Distributions/Library/wlInterpolationModule.f90
@@ -886,7 +886,7 @@ CONTAINS
     REAL(dp), INTENT(in)  :: OS
     REAL(dp), INTENT(in)  :: Table(:,:,:)
     REAL(dp), INTENT(out) :: Interpolant(:)
-    REAL(dp), INTENT(out), OPTIONAL :: Error_Option
+    INTEGER,  INTENT(out), OPTIONAL :: Error_Option
 
     INTEGER  :: &
       iP, iD, iT, iY
